rename fil for context etter navnebytte på komponenter
