+++ src/frontend/sider/Oppgavebenk/OppgavelisteNavigator.tsx
@@ -5,7 +5,7 @@ import styled from 'styled-components';
 import { HStack, Pagination } from '@navikt/ds-react';
 import { RessursStatus } from '@navikt/familie-typer';
 
-import { oppgaveSideLimit, useOppgavebenkContext } from './OppgaverContext';
+import { oppgaveSideLimit, useOppgavebenkContext } from './OppgavebenkContext';
 import type { IOppgave } from '../../typer/oppgave';
 
 const StyledSpan = styled.span`
